requirements.txt
asgiref==3.9.1
Django==4.2.10
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-json-widget==2.0.3
django-user-agents==0.4.0
djangorestframework==3.16.1
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
python-dotenv==1.1.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0